chore(storybook): js to ts
