+++ packages/wiz-ui-next/tsconfig.json
@@ -11,5 +11,11 @@
       "unplugin-vue-define-options/macros-global"
     ]
   },
-  "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.tsx", "src/**/*.vue"]
+  "include": [
+    "src/**/*.ts",
+    "src/**/*.d.ts",
+    "src/**/*.tsx",
+    "src/**/*.vue",
+    ".storybook/**/*.ts"
+  ]
 }
