+++ packages/wiz-ui-react/.storybook/main.ts
@@ -1,10 +1,10 @@
-const { vanillaExtractPlugin } = require("@vanilla-extract/vite-plugin");
-const path = require("path");
-const { mergeConfig } = require("vite");
+import type { StorybookConfig } from "@storybook/react-vite";
+import { vanillaExtractPlugin } from "@vanilla-extract/vite-plugin";
+import path from "path";
+import { mergeConfig } from "vite";
 
-module.exports = {
+const config: StorybookConfig = {
   stories: ["../src/**/*.@(mdx|stories.@(js|jsx|ts|tsx))"],
-
   addons: [
     "@storybook/addon-links",
     "@storybook/addon-essentials",
@@ -13,16 +13,10 @@ module.exports = {
     "storycap",
     "@chromatic-com/storybook",
   ],
-
   framework: {
     name: "@storybook/react-vite",
     options: {},
   },
-
-  features: {
-    interactionsDebugger: true,
-  },
-
   viteFinal: async (config) => {
     return mergeConfig(config, {
       plugins: [vanillaExtractPlugin()],
@@ -32,7 +26,6 @@ module.exports = {
       base: "./",
     });
   },
-
   staticDirs: [
     {
       from: "./assets",
@@ -48,3 +41,5 @@ module.exports = {
     reactDocgen: "react-docgen-typescript",
   },
 };
+
+export default config;
